requirements.txt
django==3.2.*
psycopg2-binary==2.9.*
tabulate-0.9.0
environs-9.5.0
